update deprecated usage in library-related plugins
