+++ lib/ExportPropertyMainTemplatePlugin.js
@@ -6,8 +6,9 @@
 "use strict";
 
 const { ConcatSource } = require("webpack-sources");
+const JavascriptModulesPlugin = require("./JavascriptModulesPlugin");
 
-/** @typedef {import("./Compilation")} Compilation */
+/** @typedef {import("./Compiler")} Compiler */
 
 /**
  * @param {string[]} accessor the accessor to convert to path
@@ -17,7 +18,7 @@ const accessorToObjectAccess = accessor => {
 	return accessor.map(a => `[${JSON.stringify(a)}]`).join("");
 };
 
-class ExportPropertyMainTemplatePlugin {
+class ExportPropertyTemplatePlugin {
 	/**
 	 * @param {string|string[]} property the name of the property to export
 	 */
@@ -26,32 +27,31 @@ class ExportPropertyMainTemplatePlugin {
 	}
 
 	/**
-	 * @param {Compilation} compilation the compilation instance
+	 * @param {Compiler} compiler the compiler instance
 	 * @returns {void}
 	 */
-	apply(compilation) {
-		const { mainTemplate, chunkTemplate } = compilation;
+	apply(compiler) {
+		compiler.hooks.thisCompilation.tap(
+			"ExportPropertyTemplatePlugin",
+			compilation => {
+				const hooks = JavascriptModulesPlugin.getCompilationHooks(compilation);
 
-		const onRenderWithEntry = (source, chunk, hash) => {
-			const postfix = accessorToObjectAccess([].concat(this.property));
-			return new ConcatSource(source, postfix);
-		};
+				hooks.renderWithEntry.tap(
+					"ExportPropertyTemplatePlugin",
+					(source, { chunk }) => {
+						const postfix = accessorToObjectAccess([].concat(this.property));
+						return new ConcatSource(source, postfix);
+					}
+				);
 
-		mainTemplate.hooks.renderWithEntry.tap(
-			"ExportPropertyMainTemplatePlugin",
-			onRenderWithEntry
+				hooks.chunkHash.tap("ExportPropertyTemplatePlugin", (chunk, hash) => {
+					if (!chunk.hasEntryModule()) return;
+					hash.update("export property");
+					hash.update(`${this.property}`);
+				});
+			}
 		);
-
-		chunkTemplate.hooks.renderWithEntry.tap(
-			"ExportPropertyMainTemplatePlugin",
-			onRenderWithEntry
-		);
-
-		mainTemplate.hooks.hash.tap("ExportPropertyMainTemplatePlugin", hash => {
-			hash.update("export property");
-			hash.update(`${this.property}`);
-		});
 	}
 }
 
-module.exports = ExportPropertyMainTemplatePlugin;
+module.exports = ExportPropertyTemplatePlugin;
