+++ lib/LibraryTemplatePlugin.js
@@ -5,9 +5,9 @@
 
 "use strict";
 
-const RuntimeGlobals = require("./RuntimeGlobals");
-const SetVarMainTemplatePlugin = require("./SetVarMainTemplatePlugin");
+const SetVarTemplatePlugin = require("./SetVarTemplatePlugin");
 
+/** @typedef {import("../declarations/WebpackOptions").LibraryCustomUmdCommentObject} LibraryCustomUmdCommentObject */
 /** @typedef {import("../declarations/WebpackOptions").LibraryCustomUmdObject} LibraryCustomUmdObject */
 /** @typedef {import("./Compiler")} Compiler */
 
@@ -48,20 +48,32 @@ const accessorAccess = (base, accessor, umdProperty, joinWith = "; ") => {
 		.join(joinWith);
 };
 
+/**
+ * @typedef {Object} LibraryOptions
+ * @property {string|string[]|LibraryCustomUmdObject=} library name of library
+ * @property {string=} libraryTarget type of library
+ * @property {boolean=} umdNamedDefine setting this to true will name the UMD module
+ * @property {string|LibraryCustomUmdCommentObject=} auxiliaryComment comment in the UMD wrapper
+ * @property {string|string[]=} libraryExport which export should be exposed as library
+ * @property {string=} globalObject global object expression
+ */
+
 class LibraryTemplatePlugin {
 	/**
-	 * @param {string|string[]|LibraryCustomUmdObject} name name of library
-	 * @param {string} target type of library
-	 * @param {boolean} umdNamedDefine setting this to true will name the UMD module
-	 * @param {string|TODO} auxiliaryComment comment in the UMD wrapper
-	 * @param {string|string[]} exportProperty which export should be exposed as library
+	 * @param {LibraryOptions} options options
 	 */
-	constructor(name, target, umdNamedDefine, auxiliaryComment, exportProperty) {
-		this.name = name;
-		this.target = target;
-		this.umdNamedDefine = umdNamedDefine;
-		this.auxiliaryComment = auxiliaryComment;
-		this.exportProperty = exportProperty;
+	constructor(options) {
+		if (arguments.length > 1) {
+			options = {
+				library: arguments[0],
+				libraryTarget: arguments[1],
+				umdNamedDefine: arguments[2],
+				auxiliaryComment: arguments[3],
+				libraryExport: arguments[4],
+				globalObject: null
+			};
+		}
+		this.options = options;
 	}
 
 	/**
@@ -69,125 +81,120 @@ class LibraryTemplatePlugin {
 	 * @returns {void}
 	 */
 	apply(compiler) {
-		compiler.hooks.thisCompilation.tap("LibraryTemplatePlugin", compilation => {
-			if (this.exportProperty) {
-				const ExportPropertyMainTemplatePlugin = require("./ExportPropertyMainTemplatePlugin");
-				new ExportPropertyMainTemplatePlugin(this.exportProperty).apply(
-					compilation
-				);
-			}
-			compilation.hooks.additionalTreeRuntimeRequirements.tap(
-				"LibraryTemplatePlugin",
-				(chunk, set) => {
-					set.add(RuntimeGlobals.returnExportsFromRuntime);
-				}
+		if (this.options.libraryExport) {
+			const ExportPropertyMainTemplatePlugin = require("./ExportPropertyMainTemplatePlugin");
+			new ExportPropertyMainTemplatePlugin(this.options.libraryExport).apply(
+				compiler
 			);
-			switch (this.target) {
-				case "var":
-					if (
-						!this.name ||
-						(typeof this.name === "object" && !Array.isArray(this.name))
-					) {
-						throw new Error(
-							"library name must be set and not an UMD custom object for non-UMD target"
-						);
-					}
-					new SetVarMainTemplatePlugin(
-						`var ${accessorAccess(undefined, this.name, "root")}`,
-						false
-					).apply(compilation);
-					break;
-				case "assign":
-					new SetVarMainTemplatePlugin(
-						accessorAccess(undefined, this.name, "root"),
-						false
-					).apply(compilation);
-					break;
-				case "this":
-				case "self":
-				case "window":
-					if (this.name) {
-						new SetVarMainTemplatePlugin(
-							accessorAccess(this.target, this.name, "root"),
-							false
-						).apply(compilation);
-					} else {
-						new SetVarMainTemplatePlugin(this.target, true).apply(compilation);
-					}
-					break;
-				case "global":
-					if (this.name) {
-						new SetVarMainTemplatePlugin(
-							accessorAccess(
-								compilation.runtimeTemplate.outputOptions.globalObject,
-								this.name,
-								"root"
-							),
-							false
-						).apply(compilation);
-					} else {
-						new SetVarMainTemplatePlugin(
-							compilation.runtimeTemplate.outputOptions.globalObject,
-							true
-						).apply(compilation);
-					}
-					break;
-				case "commonjs":
-					if (this.name) {
-						new SetVarMainTemplatePlugin(
-							accessorAccess("exports", this.name, "commonjs"),
-							false
-						).apply(compilation);
-					} else {
-						new SetVarMainTemplatePlugin("exports", true).apply(compilation);
-					}
-					break;
-				case "commonjs2":
-				case "commonjs-module":
-					new SetVarMainTemplatePlugin("module.exports", false).apply(
-						compilation
-					);
-					break;
-				case "amd":
-				case "amd-require": {
-					const AmdMainTemplatePlugin = require("./AmdMainTemplatePlugin");
-					if (this.name && typeof this.name !== "string") {
-						throw new Error("library name must be a string for amd target");
-					}
-					new AmdMainTemplatePlugin({
-						name: this.name,
-						requireAsWrapper: this.target === "amd-require"
-					}).apply(compilation);
-					break;
-				}
-				case "umd":
-				case "umd2": {
-					const UmdMainTemplatePlugin = require("./UmdMainTemplatePlugin");
-					new UmdMainTemplatePlugin(this.name, {
-						optionalAmdExternalAsGlobal: this.target === "umd2",
-						namedDefine: this.umdNamedDefine,
-						auxiliaryComment: this.auxiliaryComment
-					}).apply(compilation);
-					break;
-				}
-				case "jsonp": {
-					const JsonpExportMainTemplatePlugin = require("./web/JsonpExportMainTemplatePlugin");
-					if (typeof this.name !== "string")
-						throw new Error("library name must be a string for jsonp target");
-					new JsonpExportMainTemplatePlugin(this.name).apply(compilation);
-					break;
-				}
-				case "system": {
-					const SystemMainTemplatePlugin = require("./SystemMainTemplatePlugin");
-					new SystemMainTemplatePlugin({
-						name: this.name
-					}).apply(compilation);
-					break;
-				}
-				default:
-					throw new Error(`${this.target} is not a valid Library target`);
+		}
+		switch (this.options.libraryTarget) {
+			case "umd":
+			case "umd2": {
+				const UmdTemplatePlugin = require("./UmdTemplatePlugin");
+				new UmdTemplatePlugin(this.options.library, {
+					optionalAmdExternalAsGlobal: this.options.libraryTarget === "umd2",
+					namedDefine: this.options.umdNamedDefine,
+					auxiliaryComment: this.options.auxiliaryComment || ""
+				}).apply(compiler);
+				break;
 			}
-		});
+			case "amd":
+			case "amd-require": {
+				const AmdTemplatePlugin = require("./AmdTemplatePlugin");
+				new AmdTemplatePlugin({
+					name: this.options.library,
+					requireAsWrapper: this.options.libraryTarget === "amd-require"
+				}).apply(compiler);
+				break;
+			}
+			case "var":
+				if (
+					!this.options.library ||
+					(typeof this.options.library === "object" &&
+						!Array.isArray(this.options.library))
+				) {
+					throw new Error(
+						"library name must be set and not an UMD custom object for non-UMD target"
+					);
+				}
+				new SetVarTemplatePlugin(
+					`var ${accessorAccess(undefined, this.options.library, "root")}`,
+					false
+				).apply(compiler);
+				break;
+			case "assign":
+				new SetVarTemplatePlugin(
+					accessorAccess(undefined, this.options.library, "root"),
+					false
+				).apply(compiler);
+				break;
+			case "this":
+			case "self":
+			case "window":
+				if (this.options.library) {
+					new SetVarTemplatePlugin(
+						accessorAccess(
+							this.options.libraryTarget,
+							this.options.library,
+							"root"
+						),
+						false
+					).apply(compiler);
+				} else {
+					new SetVarTemplatePlugin(this.options.libraryTarget, true).apply(
+						compiler
+					);
+				}
+				break;
+			case "global": {
+				const globalObject =
+					this.options.globalObject || compiler.options.output.globalObject;
+				if (this.options.library) {
+					new SetVarTemplatePlugin(
+						accessorAccess(globalObject, this.options.library, "root"),
+						false
+					).apply(compiler);
+				} else {
+					new SetVarTemplatePlugin(globalObject, true).apply(compiler);
+				}
+				break;
+			}
+			case "commonjs":
+				if (this.options.library) {
+					new SetVarTemplatePlugin(
+						accessorAccess("exports", this.options.library, "commonjs"),
+						false
+					).apply(compiler);
+				} else {
+					new SetVarTemplatePlugin("exports", true).apply(compiler);
+				}
+				break;
+			case "commonjs2":
+			case "commonjs-module":
+				new SetVarTemplatePlugin("module.exports", false).apply(compiler);
+				break;
+			case "jsonp": {
+				if (!this.options.library || typeof this.options.library === "object") {
+					throw new Error(
+						"library name must be set and not an array or UMD custom object for non-UMD target"
+					);
+				}
+				const JsonpExportTemplatePlugin = require("./web/JsonpExportTemplatePlugin");
+				new JsonpExportTemplatePlugin(this.options.library).apply(compiler);
+				break;
+			}
+			case "system": {
+				const SystemTemplatePlugin = require("./SystemTemplatePlugin");
+				new SystemTemplatePlugin({
+					name: this.options.library
+				}).apply(compiler);
+				break;
+			}
+			default:
+				throw new Error(
+					`${this.options.libraryTarget} is not a valid Library target`
+				);
+		}
 	}
 }
