+++ lib/WebpackOptionsApply.js
@@ -228,13 +228,7 @@ class WebpackOptionsApply extends OptionsApply {
 
 		if (options.output.library || options.output.libraryTarget !== "var") {
 			const LibraryTemplatePlugin = require("./LibraryTemplatePlugin");
-			new LibraryTemplatePlugin(
-				options.output.library,
-				options.output.libraryTarget,
-				options.output.umdNamedDefine,
-				options.output.auxiliaryComment || "",
-				options.output.libraryExport
-			).apply(compiler);
+			new LibraryTemplatePlugin(options.output).apply(compiler);
 		}
 		if (options.externals) {
 			const ExternalsPlugin = require("./ExternalsPlugin");
