+++ lib/index.js
@@ -68,11 +68,11 @@ exportPlugins(module.exports, {
 		() => require("./EntryPlugin"),
 		"SingleEntryPlugin was renamed to EntryPlugin"
 	),
-	SetVarMainTemplatePlugin: () => require("./SetVarMainTemplatePlugin"),
+	SetVarMainTemplatePlugin: () => require("./SetVarTemplatePlugin"),
 	SourceMapDevToolPlugin: () => require("./SourceMapDevToolPlugin"),
 	Stats: () => require("./Stats"),
 	Template: () => require("./Template"),
-	UmdMainTemplatePlugin: () => require("./UmdMainTemplatePlugin"),
+	UmdMainTemplatePlugin: () => require("./UmdTemplatePlugin"),
 	WatchIgnorePlugin: () => require("./WatchIgnorePlugin")
 });
